create invoice data

+++ src/main/java/com/reports/jasper/repositories/ProductRepository.java
@@ -2,8 +2,8 @@ package com.reports.jasper.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.reports.jasper.entities.Product;
+import com.reports.jasper.domain.entities.Product;
 
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 
 }
