+++ src/main/java/com/reports/jasper/repositories/SaleOrderRepository.java
@@ -2,8 +2,8 @@ package com.reports.jasper.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.reports.jasper.entities.SaleOrder;
+import com.reports.jasper.domain.entities.SaleOrder;
 
-public interface SaleOrderRepository extends JpaRepository<SaleOrder, Long>{
+public interface SaleOrderRepository extends JpaRepository<SaleOrder, Integer>{
 
 }
